+++ src/main/kotlin/io/github/dlachouette/teamcity/github/web/HealthController.kt
@@ -27,7 +27,7 @@ class HealthController(
 
     init {
         webManager.registerController(PATH, this)
-        authInterceptor.addPathNotRequiringAuth(PATH)
+        authInterceptor.addPathNotRequiringAuth(javaClass, PATH)
     }
 
     override fun doHandle(request: HttpServletRequest, response: HttpServletResponse): ModelAndView? {